+++ cypress/e2e/MapControls.cy.js
@@ -12,12 +12,7 @@ describe("MapControls", () => {
       .then(($mapControlsEl) => window.getComputedStyle($mapControlsEl[0]))
       .invoke("getPropertyValue", "right")
       .should("equal", "0px");
-    cy.get(".wkp-search-input input")
-      .focus()
-      .type("B")
-      .type("e")
-      .type("r")
-      .type("n");
+    cy.get(".wkp-search-input input").focus().type("Bern");
     cy.get("#react-autowhatever-1-section-0-item-0").click({ force: true });
     const overlay = cy.get(".wkp-feature-information", { timeout: 40000 });
     overlay.should("be.visible");