+++ src/popups/StopPlacePopup/StopPlacePopup.test.js
@@ -92,7 +92,7 @@ describe("StopPlacePopup", () => {
     expect(queryByTestId("stopplace-url")).toBeFalsy();
   });
 
-  test("only displays alternative-transport box with PARTIALLY and with note", async () => {
+  test("only displays alternative-transport box with only note when state = PARTIALLY", async () => {
     const note =
       "Very important information for people with disabilities for alternative transport";
     const spy = fetchMock.once(new RegExp(cartaroUrl, "g"), {
@@ -116,14 +116,14 @@ describe("StopPlacePopup", () => {
 
     expect(queryByTestId("stopplace-accessibility")).toBeFalsy();
     expect(queryByTestId("stopplace-alternative-transport")).toBeTruthy();
-    expect(queryByText("Teilweise")).toBeTruthy();
+    expect(queryByText("Teilweise")).toBeFalsy();
     expect(queryByText(note)).toBeTruthy();
     expect(queryByTestId("stopplace-passengerinfo")).toBeFalsy();
     expect(queryByTestId("stopplace-note")).toBeFalsy();
     expect(queryByTestId("stopplace-url")).toBeFalsy();
   });
 
-  test.only("only displays alternative-transport box with Ja", async () => {
+  test("only displays alternative-transport box with Shuttle-Fahrdienst when state = YES and no note is defined", async () => {
     const spy = fetchMock.once(new RegExp(cartaroUrl, "g"), {
       prmInformation: {
         alternativeTransport: {
@@ -144,14 +144,52 @@ describe("StopPlacePopup", () => {
 
     expect(queryByTestId("stopplace-accessibility")).toBeFalsy();
     expect(queryByTestId("stopplace-alternative-transport")).toBeTruthy();
-    expect(queryByText("Ja")).toBeTruthy();
+    expect(queryByTestId("stopplace-alternative-transport-state")).toBeTruthy();
+    expect(queryByText("Shuttle-Fahrdienst")).toBeTruthy();
     expect(queryByTestId("stopplace-passengerinfo")).toBeFalsy();
     expect(queryByTestId("stopplace-note")).toBeFalsy();
     expect(queryByTestId("stopplace-url")).toBeFalsy();
   });
 
-  ["NO", "UNKNOWN"].forEach((state) => {
-    test(`does not display alternative-transport box when value is ${state}`, async () => {
+  ["NO", "UNKNOWN", "NOT_APPLICABLE"].forEach((state) => {
+    test(`displays alternative-transport box when value is ${state} but there is a note`, async () => {
+      const note = "Very important information for people with disabilities";
+      const spy = fetchMock.once(new RegExp(cartaroUrl, "g"), {
+        prmInformation: {
+          alternativeTransport: {
+            state,
+            note,
+          },
+        },
+      });
+
+      const { queryByTestId, queryByText } = render(
+        <Provider store={store}>
+          <StopPlacePopup feature={new Feature({ uic: `show-${state}` })} />
+        </Provider>,
+      );
+
+      await waitFor(() => {
+        return spy.called();
+      });
+
+      expect(queryByTestId("stopplace-accessibility")).toBeFalsy();
+      expect(queryByTestId("stopplace-alternative-transport")).toBeTruthy();
+      expect(
+        queryByTestId("stopplace-alternative-transport-state"),
+      ).toBeFalsy();
+      expect(
+        queryByTestId("stopplace-alternative-transport-note"),
+      ).toBeTruthy();
+      expect(queryByText(note)).toBeTruthy();
+      expect(queryByTestId("stopplace-passengerinfo")).toBeFalsy();
+      expect(queryByTestId("stopplace-note")).toBeFalsy();
+      expect(queryByTestId("stopplace-url")).toBeFalsy();
+    });
+  });
+
+  ["NO", "UNKNOWN", "NOT_APPLICABLE", "PARTIALLY"].forEach((state) => {
+    test(`displays no alternative-transport box when value is ${state} without note`, async () => {
       const spy = fetchMock.once(new RegExp(cartaroUrl, "g"), {
         prmInformation: {
           alternativeTransport: {
@@ -162,7 +200,7 @@ describe("StopPlacePopup", () => {
 
       const { queryByTestId } = render(
         <Provider store={store}>
-          <StopPlacePopup feature={new Feature({ uic: state })} />
+          <StopPlacePopup feature={new Feature({ uic: `hide-${state}` })} />
         </Provider>,
       );
 
@@ -172,6 +210,10 @@ describe("StopPlacePopup", () => {
 
       expect(queryByTestId("stopplace-accessibility")).toBeFalsy();
       expect(queryByTestId("stopplace-alternative-transport")).toBeFalsy();
+      expect(
+        queryByTestId("stopplace-alternative-transport-state"),
+      ).toBeFalsy();
+      expect(queryByTestId("stopplace-alternative-transport-note")).toBeFalsy();
       expect(queryByTestId("stopplace-passengerinfo")).toBeFalsy();
       expect(queryByTestId("stopplace-note")).toBeFalsy();
       expect(queryByTestId("stopplace-url")).toBeFalsy();